Remove appveyor logging as its not possible to make it work perfectly
For NUnit tests to be logged in appveyor you need to the use the AppVeyor logger. 
Unfortunately since we run the same test twice, once for .net core and once for .net framework the AppVeyor logger doesn't detect the difference and only logs the test once, which results in incorrect numbers. 

Removed AppVeyor  logger fully and not going to rely on the number in CI.