Updated Snapper to create snapshot if none exists (#71)
* Updated Snapper to create snapshot if none exists

* Update test projects to .net core 3.1

* Bump appveyor image to VS 2019

* Fix bug with updating snapshots in CI env

* Update docs to mention auto creating of initial snapshot

* Update test projects to use .net 452 so that the latest xunit works

* Remove appveyor logging as its not possible to make it work perfectly

For NUnit tests to be logged in appveyor you need to the use the AppVeyor logger. 
Unfortunately since we run the same test twice, once for .net core and once for .net framework the AppVeyor logger doesn't detect the difference and only logs the test once, which results in incorrect numbers. 

Removed AppVeyor  logger fully and not going to rely on the number in CI.